lazy load client dropzone
