+++ src/db/queries.ts
@@ -16,19 +16,3 @@ export const getImagesForUser = async () => {
 type Awaited<T> = T extends PromiseLike<infer U> ? Awaited<U> : T;
 
 export type ImageFromDb = Awaited<ReturnType<typeof getImagesForUser>>[0];
-
-export const getGroupedImagesForUser = async () => {
-  const images = await getImagesForUser();
-
-  const imagesByDate = images.reduce(
-    (acc: Record<string, ImageFromDb[]>, image) => {
-      const date = image.createdAt.toDateString();
-      if (!acc[date]) acc[date] = [];
-      acc[date].push(image);
-      return acc;
-    },
-    {}
-  );
-
-  return imagesByDate;
-};
